[Redux] Move category states from queriesSlice to budgetSlice
