+++ packages/desktop-client/src/components/mobile/budget/BudgetPage.tsx
@@ -50,7 +50,7 @@ import {
   deleteCategoryGroup,
   updateCategory,
   updateCategoryGroup,
-} from '@desktop-client/queries/queriesSlice';
+} from '@desktop-client/budget/budgetSlice';
 import { useDispatch } from '@desktop-client/redux';
 import {
   envelopeBudget,
