+++ packages/desktop-client/src/components/modals/manager/DeleteFileModal.tsx
@@ -6,7 +6,7 @@ import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 
-import { deleteBudget } from '@desktop-client/budgets/budgetsSlice';
+import { deleteBudget } from '@desktop-client/budgetfiles/budgetfilesSlice';
 import {
   Modal,
   ModalCloseButton,
