+++ packages/desktop-client/src/components/modals/manager/FilesSettingsModal.tsx
@@ -8,7 +8,7 @@ import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 
-import { loadAllFiles } from '@desktop-client/budgets/budgetsSlice';
+import { loadAllFiles } from '@desktop-client/budgetfiles/budgetfilesSlice';
 import {
   Modal,
   ModalCloseButton,
