+++ packages/desktop-client/src/components/modals/manager/ImportYNAB5Modal.tsx
@@ -9,7 +9,7 @@ import { styles } from '@actual-app/components/styles';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 
-import { importBudget } from '@desktop-client/budgets/budgetsSlice';
+import { importBudget } from '@desktop-client/budgetfiles/budgetfilesSlice';
 import { Link } from '@desktop-client/components/common/Link';
 import {
   Modal,
