+++ packages/desktop-client/src/global-events.ts
@@ -4,7 +4,8 @@ import * as undo from 'loot-core/platform/client/undo';
 
 import { reloadAccounts } from './accounts/accountsSlice';
 import { setAppState } from './app/appSlice';
-import { closeBudgetUI } from './budgets/budgetsSlice';
+import { reloadCategories } from './budget/budgetSlice';
+import { closeBudgetUI } from './budgetfiles/budgetfilesSlice';
 import { closeModal, pushModal, replaceModal } from './modals/modalsSlice';
 import {
   addGenericErrorNotification,
@@ -12,7 +13,6 @@ import {
 } from './notifications/notificationsSlice';
 import { reloadPayees } from './payees/payeesSlice';
 import { loadPrefs } from './prefs/prefsSlice';
-import { reloadCategories } from './queries/queriesSlice';
 import { type AppStore } from './redux/store';
 import * as syncEvents from './sync-events';
