+++ packages/desktop-client/src/index.tsx
@@ -19,7 +19,7 @@ import { q } from 'loot-core/shared/query';
 import * as accountsSlice from './accounts/accountsSlice';
 import * as appSlice from './app/appSlice';
 import { AuthProvider } from './auth/AuthProvider';
-import * as budgetsSlice from './budgets/budgetsSlice';
+import * as budgetfilesSlice from './budgetfiles/budgetfilesSlice';
 // See https://github.com/WICG/focus-visible. Only makes the blue
 // focus outline appear from keyboard events.
 import 'focus-visible';
@@ -38,7 +38,7 @@ const boundActions = bindActionCreators(
   {
     ...accountsSlice.actions,
     ...appSlice.actions,
-    ...budgetsSlice.actions,
+    ...budgetfilesSlice.actions,
     ...modalsSlice.actions,
     ...notificationsSlice.actions,
     ...prefsSlice.actions,
