+++ packages/desktop-client/src/sync-events.ts
@@ -5,7 +5,7 @@ import { listen, send } from 'loot-core/platform/client/fetch';
 
 import { reloadAccounts } from './accounts/accountsSlice';
 import { resetSync, sync } from './app/appSlice';
-import { closeAndDownloadBudget, uploadBudget } from './budgets/budgetsSlice';
+import { closeAndDownloadBudget, uploadBudget } from './budgetfiles/budgetfilesSlice';
 import { pushModal } from './modals/modalsSlice';
 import {
   addNotification,
@@ -13,7 +13,7 @@ import {
 } from './notifications/notificationsSlice';
 import { reloadPayees } from './payees/payeesSlice';
 import { loadPrefs } from './prefs/prefsSlice';
-import { reloadCategories } from './queries/queriesSlice';
+import { reloadCategories } from './budget/budgetSlice';
 import { type AppStore } from './redux/store';
 import { signOut } from './users/usersSlice';
