+++ packages/loot-core/src/server/budget/app.ts
@@ -316,7 +316,7 @@ async function moveCategory({
 }: {
   id: CategoryEntity['id'];
   groupId: CategoryGroupEntity['id'];
-  targetId: CategoryEntity['id'];
+  targetId: CategoryEntity['id'] | null;
 }): Promise<void> {
   await batchMessages(async () => {
     await db.moveCategory(id, groupId, targetId);
@@ -328,7 +328,7 @@ async function deleteCategory({
   transferId,
 }: {
   id: CategoryEntity['id'];
-  transferId?: CategoryEntity['id'];
+  transferId?: CategoryEntity['id'] | null;
 }): Promise<{ error?: 'no-categories' | 'category-type' }> {
   let result = {};
   await batchMessages(async () => {
@@ -407,7 +407,7 @@ async function moveCategoryGroup({
   targetId,
 }: {
   id: CategoryGroupEntity['id'];
-  targetId: CategoryGroupEntity['id'];
+  targetId: CategoryGroupEntity['id'] | null;
 }): Promise<void> {
   await batchMessages(async () => {
     await db.moveCategoryGroup(id, targetId);
@@ -419,7 +419,7 @@ async function deleteCategoryGroup({
   transferId,
 }: {
   id: CategoryGroupEntity['id'];
-  transferId: CategoryGroupEntity['id'];
+  transferId?: CategoryGroupEntity['id'] | null;
 }): Promise<void> {
   const groupCategories = await db.all<Pick<CategoryEntity, 'id'>>(
     'SELECT id FROM categories WHERE cat_group = ? AND tombstone = 0',
