+++ upcoming-release-notes/5593.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [joel-jeremy]
+---
+
+[Redux] Move category states from queriesSlice to budgetSlice
